Tests for API inputs.
Added tests for several API endpoints to assist in preventing XSS
vulnerabilities.
